:hammer: Package reducers/actions/selectors into bundes
