+++ src/state/recipes/doFetchRecipes.js
@@ -1,7 +1,4 @@
 import fetch from 'isomorphic-fetch';
-import { compose } from 'redux';
-import { selectRecipes } from '../../selectors/recipes';
-import { shouldFetchMetadata } from '../metadata';
 
 const API_URL = 'https://api.punkapi.com/v2/beers';
 
@@ -14,16 +11,5 @@ export const doFetchRecipes = () => async dispatch => {
 	}
 	catch (error) {
 		dispatch({ type: 'FETCH_RECIPES_FAILED', payload: error });
-	}
-};
-
-const shouldFetchRecipes = compose(
-	shouldFetchMetadata({ cacheDuration: 60 * 1000 }),
-	selectRecipes
-);
-
-export const doFetchRecipesIfNeeded = () => async (dispatch, getState) => {
-	if (shouldFetchRecipes(getState())) {
-		return dispatch(doFetchRecipes());
 	}
 };
