+++ src/state/recipes/recipesReducer.js
@@ -1,0 +1,3 @@
+import { getMetadataReducer } from '../metadata/getMetadataReducer';
+
+export const recipesReducer = getMetadataReducer({ baseType: 'FETCH_RECIPES' });
