+++ src/state/rootReducer.js
@@ -1,8 +1,8 @@
 import { routerReducer } from 'preact-router-redux';
 import { combineReducers } from 'redux';
-import { getMetadataReducer } from './metadata';
+import { recipesReducer } from './recipes/recipesReducer';
 
 export const rootReducer = combineReducers({
 	routing: routerReducer,
-	recipes: getMetadataReducer({ baseType: 'FETCH_RECIPES' })
+	recipes: recipesReducer
 });
